Create Pattern count
PatternCount is just one of many functions that we will encounter in this book. Sometimes, we will ask you to implement these functions in "Code Challenges". We have already done the first exercise for you (i.e., all you need to do is copy the function to the box below). This step is designed to help get you used to the environment.

def PatternCount(Text, Pattern):
    count = 0
    for i in range(len(Text)-len(Pattern)+1):
        if Text[i:i+len(Pattern)] == Pattern:
            count = count+1
    return count 
So to get started, click the "Click to Start Solving" button below. Then, copy the PatternCount code provided into the designated space in the window below and click "Submit".  As with Exercise Breaks, don't worry if you don't solve the problem correctly on the first try; you can take as many attempts as you like. Later in the chapter, we will take off the training wheels and let you start writing your own functions.

Some notes on how code challenges work:

We ask you to write a function on each problem with the specified input and output.  Here, your function should be called PatternCount() and take strings Text and Pattern as input parameters (in that order).  If you need any subroutines, provide those as well.
Behind the scenes, we grade your code on a collection of test datasets.  You either pass all the tests (congrats!) or we let you know the first test that your code failed on.  We provide a link to test datasets beneath each problem so that you can test your code before submitting your function. Click here for these datasets for this step.
We also provide a "sample input and output" so that you can see a single example of the correct answer for a given sample input.
If your problem is graded incorrect but you don't get feedback on your code, please tell us what your mistake was so that we can improve the problem for future students. You can report your issue in the discussion forum at the bottom of the code challenge page.
Sample Input:

GCGCG
GCG
Sample Output:

2