Create Translating RNA to proteins
Protein translation follows the universal code, where each 3 bases (1 codon) in mRNA represent one amino acid (aa). The genetic code is unique for each aa. In other words, no aa shares genetic code with another. Also, genetic code covers all possibilities of 3 consecutive bases (4^3), but there are still 3 codons not equivalent to any aa, but rather to a stop signal.
We have a dictionary defining the genetic code, with a string containing all aa. We slide a window along the RNA sequence (with a length equals 1 codon/3 bases). While sliding, we search for each window in the lists of genetic code (values of the dict) and get the key (the correct aa symbol) that contains such window, and concatenate it to the peptide string.